Refactor builds to avoid webpack CLI.
