+++ src/test/mochitest/examples/doc-sourcemapped.html
@@ -10,7 +10,7 @@
   <body>
     <!-- INJECTED-START -->
     <!--
-      Content generated by examples/sourcemapped/webpack.config.js.
+      Content generated by examples/sourcemapped/build.js.
       Run "yarn build" to update.
     -->
     <script src="sourcemapped/fixtures/babel-bindings-with-flow/output.js"></script>
